Fixed various errors that occur if the appdata sub-directory (Config.ProgramDataPath) doesn't exist. Wasn't a problem for proper installation, because the installer creates the directory. Was a problem if the installation directory was copied to a machine where the installer was never run.

+++ compLexity Demo Player/FileOperationList.cs
@@ -41,7 +41,12 @@ namespace compLexity_Demo_Player
             }
 
             operations.Clear();
-            File.Delete(programPath + "\\" + fileName);
+            String configFullPath = programPath + "\\" + fileName;
+
+            if (File.Exists(configFullPath))
+            {
+                File.Delete(programPath + "\\" + fileName);
+            }
         }
 
         static public void Initialise(String _programPath, String _fileName)
